🚚 Move version list component to its own folder
